LayoutTests/http/wpt/offscreen-canvas/transferToImageBitmap-empty.html

    <!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="help" href="https://html.spec.whatwg.org/#dom-offscreencanvas">
<script>

test(function() {
    let offscreenCanvas = new OffscreenCanvas(100, 100);
    assert_throws_dom('InvalidStateError', () => { offscreenCanvas.transferToImageBitmap() });
}, "Test transferToImageBitmap.");

</script>
